Refactoring theme provider
