+++ src/components/Layout/Layout.test.jsx
@@ -3,8 +3,8 @@ import { render, screen } from '@testing-library/react';
 import { createMemoryHistory } from 'history';
 import { Router } from 'react-router';
 import Layout from './Layout.component';
-import Theme from '../App/App.styled';
 import { SearchProvider } from '../../Hooks/SearchProvider/SearchProvider';
+import { Theme } from '../../Hooks/Theme/Theme';
 
 describe('Layout Component Tests', () => {
   const home = '/';
