+++ src/components/Menu/Menu.component.jsx
@@ -1,8 +1,7 @@
 import React from 'react';
-import { IconButton, Typography, Button } from '@material-ui/core';
+import { Typography } from '@material-ui/core';
 import YouTubeIcon from '@material-ui/icons/YouTube';
 import { useHistory, useLocation } from 'react-router-dom';
-import Brightness4Icon from '@material-ui/icons/Brightness4';
 import {
   CustomAppBar,
   CustomButton,
@@ -16,7 +15,7 @@ import {
   ToolbarSection,
 } from './Menu.styled';
 import { useSearchDispatch } from '../../Hooks/SearchProvider/SearchProvider';
-import { useThemeDispatch } from '../../Hooks/Theme/Theme';
+import { useSwitchTheme } from '../../Hooks/Theme/Theme';
 
 function HomeButton() {
   const history = useHistory();
@@ -34,43 +33,40 @@ function HomeButton() {
   );
 }
 
-function SearchBar() {
+function SearchBar(path) {
   const setSearch = useSearchDispatch();
+  const homePage = '/';
   return (
-    <SearchContainer data-testid="SearchBar">
-      <SearchIconContainer>
-        <SearchIcon />
-      </SearchIconContainer>
-      <CustomInputBase
-        data-testid="SearchInput"
-        placeholder="Search…"
-        onChange={(e) => {
-          setSearch(e.target.value);
-        }}
-      />
-    </SearchContainer>
+    path === homePage && (
+      <SearchContainer data-testid="SearchBar">
+        <SearchIconContainer>
+          <SearchIcon />
+        </SearchIconContainer>
+        <CustomInputBase
+          data-testid="SearchInput"
+          placeholder="Search…"
+          onChange={(e) => {
+            setSearch(e.target.value);
+          }}
+        />
+      </SearchContainer>
+    )
   );
 }
 
 function Menu() {
-  const [checked, setChecked] = React.useState(false);
-  const setTheme = useThemeDispatch();
-  const handleChange = () => {
-    setChecked(!checked);
-    setTheme(checked ? 'light' : 'dark');
-  };
+  const switchTheme = useSwitchTheme();
   const location = useLocation();
-
   return (
     <>
       <CustomAppBar position="fixed" data-testid="CustomAppBar">
         <CustomToolbar>
           <ToolbarSection>
             {HomeButton()}
-            {location.pathname === '/' && SearchBar()}
+            {SearchBar(location.pathname)}
           </ToolbarSection>
           <ToolbarSection>
-            <ThemeIcon fontSize="large" onClick={handleChange} />
+            <ThemeIcon fontSize="default" onClick={switchTheme} />
             <CustomButton>
               <Typography variant="body1" noWrap>
                 Login
