+++ src/components/Menu/Menu.test.jsx
@@ -3,8 +3,8 @@ import { render, screen } from '@testing-library/react';
 import { Router } from 'react-router-dom';
 import { createMemoryHistory } from 'history';
 import Menu from './Menu.component';
-import Theme from '../App/App.styled';
 import { SearchProvider } from '../../Hooks/SearchProvider/SearchProvider';
+import { Theme } from '../../Hooks/Theme/Theme';
 
 describe('Menu Component Tests', () => {
   it('Should render menu with search bar', async () => {
@@ -31,31 +31,13 @@ describe('Menu Component Tests', () => {
     render(
       <Router history={history}>
         <Theme>
-          <Menu />
+          <SearchProvider>
+            <Menu />
+          </SearchProvider>
         </Theme>
       </Router>
     );
     expect(screen.getByTestId('CustomAppBar')).toBeInTheDocument();
     expect(screen.queryByTestId('SearchBar')).not.toBeInTheDocument();
-  });
-
-  it('Should change the search bar theme switch state', async () => {
-    const anythingElseButHome = '*';
-    const history = createMemoryHistory();
-    history.push(anythingElseButHome);
-    render(
-      <Router history={history}>
-        <Theme>
-          <Menu />
-        </Theme>
-      </Router>
-    );
-    const themeSwitch = screen
-      .getByTestId('ThemeSwitch')
-      .getElementsByTagName('input')[0];
-    themeSwitch.click();
-    expect(themeSwitch.checked).toEqual(true);
-    themeSwitch.click();
-    expect(themeSwitch.checked).toEqual(false);
   });
 });
