+++ src/pages/Home/Home.page.test.jsx
@@ -2,9 +2,9 @@ import React from 'react';
 import { createMemoryHistory } from 'history';
 import { Router } from 'react-router';
 import { render, screen } from '@testing-library/react';
-import Theme from '../../components/App/App.styled';
 import HomePage from './Home.page';
 import { SearchProvider } from '../../Hooks/SearchProvider/SearchProvider';
+import { Theme } from '../../Hooks/Theme/Theme';
 
 jest.mock('../../Hooks/Video/Video');
 describe('Home View Tests', () => {
@@ -14,11 +14,13 @@ describe('Home View Tests', () => {
   beforeEach(() => {
     render(
       <Theme>
-        <Router history={history}>
-          <SearchProvider>
-            <HomePage />
-          </SearchProvider>
-        </Router>
+        <Theme>
+          <Router history={history}>
+            <SearchProvider>
+              <HomePage />
+            </SearchProvider>
+          </Router>
+        </Theme>
       </Theme>
     );
   });
