+++ src/pages/NotFound/NotFound.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 import NotFoundPage from './NotFound.page';
-import Theme from '../../components/App/App.styled';
+import { Theme } from '../../Hooks/Theme/Theme';
 
 describe('Not Found Landing View', () => {
   beforeEach(() => {
